(#25) refactor: 네비게이션 사용하는 곳에서 쓸 컴퍼넌트 추가
